High Score Page
Brings the admin to a level select screen where they could select which
level he wants to view the scores for. When you click on the level,
nothing happens. Once we implement scores then we could bring the
person to a page with the list of scores for that level.